% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{chrToNum}
\alias{chrToNum}
\title{Convert protein sequence into numeric form}
\usage{
chrToNum(protVec)
}
\arguments{
\item{protVec}{Chr, a character vector of amino acid letters}
}
\value{
Numerical, A numerical vector
}
\description{
A helper function that converts amino acid letters ("A, G, I, L, P, V, F, W, Y, D,
E, R, H, K, S, T, C, M, N, Q") into numbers in that order starting from 1
}
\examples{
tempVec <- c("A", "C", "G")
numericForm <- mutVis:::chrToNum(tempVec)
numericForm

}
