% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{compareAndUpdate}
\alias{compareAndUpdate}
\title{compare numeric and update dataframe}
\usage{
compareAndUpdate(unknown, known, protFrame, knownCount)
}
\arguments{
\item{unknown}{Numerical, A numerical vector of the unknown protein sequence}

\item{known}{Numerical, A numerical vector of the known protein sequence}

\item{protFrame}{The data frame that needs to be updated}

\item{knownCount}{Numerical, the number of known proteins that has been compared}
}
\value{
An undated version of input dataframe
}
\description{
A function that modifies the input data frame
}
