% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{emptyDF}
\alias{emptyDF}
\title{Create empty dataframe}
\usage{
emptyDF(protVector)
}
\arguments{
\item{protVector}{Chr, a character vector}
}
\value{
a dataframe with length(protVector) + 1 columns and 0 rows
}
\description{
A function that creates an empty dataframe (one character per column)
}
