% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{findPercentage}
\alias{findPercentage}
\title{Find percentage of matching amino acids}
\usage{
findPercentage(protSeq, knownProt)
}
\arguments{
\item{protSeq}{Chr, A character vector of unknown protein sequence}

\item{knownProt}{Chr, A character vector of a known protein sequence}
}
\value{
Numerical, The percentage of TRUE after comparing or 0 if the vectors are of different lengths
}
\description{
A function that calculates the percentage of TRUE after comparing two vectors
}
