% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{matched}
\alias{matched}
\title{Find best match proteins with unknown protein}
\usage{
matched(protSeq, dataset = protData, minMatch)
}
\arguments{
\item{protSeq}{Chr, The character vector of the input protein sequence}

\item{dataset}{The dataset in .fasta format that contains the protein sequences of known proteins and their names}

\item{minMatch}{Numerical, the minimum percent of matching}
}
\value{
A dataframe specifying where the input protein differs from the known proteins
}
\description{
A function that generates a dataframe listing the amino acid differences in
numeric form between the input and the known proteins
}
\examples{
inputProt <- paste("MAVLILVLLAVVILQAAPIRKLEDLLPTRYPPDHELVYWCTYANQCDFCWECVH",
"GICRNRIQADWPVIHQNDWIINCTVSRWNGICSYYEGPRNHTDHQMDCANPTSHTYPHREYMKIYERDDL",
sep = "")
unknown1 <- seqinr::s2c(inputProt)
minMatch <- 98
result <- matched(unknown1, protData, minMatch)
result

}
