% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchProt.R
\name{matchProt}
\alias{matchProt}
\title{Find the most plausible protein sequence matches of unknown input and plot
the amino acid differences}
\usage{
matchProt(unknownProt, dataset = protData, minMatch = 85)
}
\arguments{
\item{unknownProt}{String, The unknown protein sequence.}

\item{dataset}{The dataset that stores the information of known proteins with two columns, name and sequence.}

\item{minMatch}{The minimum percentage of same amino acid with unknown (default is 85\%)}
}
\value{
A graph of all the amino acid differences between input and known proteins, with the position on input protein
}
\description{
A function that lists the best matched protein sequences, compares the
candidate sequence with the unknown input, and plots the differences
}
\examples{
unknownProt <- paste("MAVLILVLLAVVILQAAPIRKLEDLLPTRYPPDHELVYWCTYANQCDFCWECVH",
"GICRNRIQADWPVIHQNDWIINCTVSRWNGICSYYEGPRNHTDHQMDCANPTSHTYPHREYMKIYERDDL",
sep = "")
minPercent <- 85
result <- matchProt(unknownProt, minMatch = minPercent)
result

}
