% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protData}
\alias{protData}
\title{protData}
\format{
A list generated from seqinr::read.fasta(file = "uniprot_sprot.fasta", seqtype = "AA") containing the known protein information:
\describe{
\item{name}{protein name}
\item{SeqFastaAA}{the sequence of the protein and information}

...
}
}
\source{
\url{http://uniprot.org}
}
\usage{
protData
}
\description{
A dataset containing 29 proteins used as comparisons
}
\examples{
\dontrun{
 protData
}

}
\keyword{datasets}
