% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMutVisApp.R
\name{runMutVisApp}
\alias{runMutVisApp}
\title{Launch the shiny app for mutVis Package}
\usage{
runMutVisApp()
}
\value{
Does not return anything. It opens a shiny page
}
\description{
A function that launches the shiny app for muVis package.
The code can be found in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runMutVisApp()
}

}
